+++ src/bit/bit.h
@@ -6,12 +6,12 @@
 class Bit : public Paintable
 {
 public:
-  Bit(double height, double width, const Vector2D &position, const Color &color);
+  Bit(float height, float width, const Vector2D &position, const Color &color);
   void paint() const override;
 
 private:
-  double paintableHeight;
-  double paintableWidth;
+  float paintableHeight;
+  float paintableWidth;
 };
 
 #endif // BIT_H
