+++ src/bpmLabel/bpmLabel.h
@@ -6,13 +6,13 @@
 class BpmLabel : public Paintable
 {
 public:
-  BpmLabel(double height, double width, const Vector2D &position);
+  BpmLabel(float height, float width, const Vector2D &position);
   void paint() const override;
   int getValue();
 
 private:
   int value = 120;
-  void paintSquares(const Vector2D &pos, double side) const;
+  void paintSquares(const Vector2D &pos, float side) const;
 };
 
 #endif // BIT_H
